+++ Magick++/lib/Magick++/Image.h
@@ -814,13 +814,13 @@ namespace Magick
     // values.  A typical strokeDashArray_ array might contain the
     // members 5 3 2 0, where the zero value indicates the end of the
     // pattern array.
-    void            strokeDashArray ( const unsigned int* strokeDashArray_ );
-    const unsigned int* strokeDashArray ( void ) const;
+    void            strokeDashArray ( const double* strokeDashArray_ );
+    const double* strokeDashArray ( void ) const;
 
     // While drawing using a dash pattern, specify distance into the
     // dash pattern to start the dash (default 0).
-    void            strokeDashOffset ( unsigned int strokeDashOffset_ );
-    unsigned int    strokeDashOffset ( void ) const;
+    void            strokeDashOffset ( double strokeDashOffset_ );
+    double    strokeDashOffset ( void ) const;
 
     // Specify the shape to be used at the end of open subpaths when
     // they are stroked. Values of LineCap are UndefinedCap, ButtCap,
