+++ Magick++/lib/Magick++/Options.h
@@ -163,13 +163,13 @@ namespace Magick
     // specify the lengths of alternating dashes and gaps in user
     // units. If an odd number of values is provided, then the list of
     // values is repeated to yield an even number of values.
-    void            strokeDashArray ( const unsigned int* strokeDashArray_ );
-    const unsigned int* strokeDashArray ( void ) const;
+    void            strokeDashArray ( const double* strokeDashArray_ );
+    const double* strokeDashArray ( void ) const;
 
     // While drawing using strokeDashArray, specify distance into the dash
     // pattern to start the dash (default 0).
-    void            strokeDashOffset ( unsigned int strokdDashOffset_ );
-    unsigned int    strokeDashOffset ( void ) const;
+    void            strokeDashOffset ( double strokdDashOffset_ );
+    double    strokeDashOffset ( void ) const;
 
     // Specify the shape to be used at the end of open subpaths when
     // they are stroked. Values of LineCap are UndefinedCap, ButtCap,
@@ -301,11 +301,11 @@ inline Magick::CompressionType Magick::Options::compressType ( void ) const
   return static_cast<Magick::CompressionType>(_imageInfo->compression);
 }
 
-inline void Magick::Options::strokeDashOffset ( unsigned int strokeDashOffset_ )
+inline void Magick::Options::strokeDashOffset ( double strokeDashOffset_ )
 {
   _drawInfo->dash_offset = strokeDashOffset_;
 }
-inline unsigned int Magick::Options::strokeDashOffset ( void ) const
+inline double Magick::Options::strokeDashOffset ( void ) const
 {
   return _drawInfo->dash_offset;
 }
